+++ DisCatSharp.ApplicationCommands/Context/ApplicationCommandsTranslationResolver.cs
@@ -25,16 +25,27 @@ using DisCatSharp.Entities;
 
 namespace DisCatSharp.ApplicationCommands
 {
+    /// <summary>
+    /// Represents a application commands translation resolver.
+    /// </summary>
     public class ApplicationCommandsTranslationResolver
     {
         #pragma warning disable IDE1006
-
+        /// <summary>
+        /// Gets the group translations.
+        /// </summary>
         public IReadOnlyDictionary<string, GroupTranslation> GroupTranslations => this._groupTranslations;
         private Dictionary<string, GroupTranslation> _groupTranslations => new();
 
+        /// <summary>
+        /// Gets the sub group translations.
+        /// </summary>
         public IReadOnlyDictionary<string, SubGroupTranslation> SubGroupTranslations => this._subGroupTranslations;
         private Dictionary<string, SubGroupTranslation> _subGroupTranslations => new();
 
+        /// <summary>
+        /// Gets the command translations.
+        /// </summary>
         public IReadOnlyDictionary<string, CommandTranslation> CommandTranslations => this._commandTranslations;
         private Dictionary<string, CommandTranslation> _commandTranslations => new();
 
@@ -43,55 +54,88 @@ namespace DisCatSharp.ApplicationCommands
 
     public class GroupTranslation
     {
+        /// <summary>
+        /// Gets the group name.
+        /// </summary>
         public string GroupName { get; set; }
 
         public DiscordApplicationCommandLocalization GroupNameLocalizations { get; set; }
 
         public DiscordApplicationCommandLocalization GroupDescriptionLocalizations { get; set; }
 
+        /// <summary>
+        /// Gets the sub group translations.
+        /// </summary>
         public Dictionary<string, SubGroupTranslation> SubGroupTranslations { get; set; }
 
+        /// <summary>
+        /// Gets the command translations.
+        /// </summary>
         public Dictionary<string, CommandTranslation> CommandTranslations { get; set; }
     }
 
     public class SubGroupTranslation
     {
+        /// <summary>
+        /// Gets the sub group name.
+        /// </summary>
         public string SubGroupName { get; set; }
 
         public DiscordApplicationCommandLocalization SubGroupNameLocalizations { get; set; }
 
         public DiscordApplicationCommandLocalization SubGroupDescriptionLocalizations { get; set; }
 
+        /// <summary>
+        /// Gets the command translations.
+        /// </summary>
         public Dictionary<string, CommandTranslation> CommandTranslations { get; set; }
     }
 
     public class CommandTranslation
     {
+        /// <summary>
+        /// Gets the command name.
+        /// </summary>
         public string CommandName { get; set; }
 
         public DiscordApplicationCommandLocalization CommandNameLocalizations { get; set; }
 
         public DiscordApplicationCommandLocalization CommandDescriptionLocalizations { get; set; }
 
-        public Dictionary<string, OptionTranslation> OptionsTranslations { get; set; }
+        /// <summary>
+        /// Gets the option translations.
+        /// </summary>
+        public Dictionary<string, OptionTranslation> OptionTranslations { get; set; }
     }
 
     public class OptionTranslation
     {
+        /// <summary>
+        /// Gets the option name.
+        /// </summary>
         public string OptionName { get; set; }
 
         public DiscordApplicationCommandLocalization OptionNameLocalizations { get; set; }
 
         public DiscordApplicationCommandLocalization OptionDescriptionLocalizations { get; set; }
 
-        public Dictionary<string, ChoiceTranslation> ChoicesTranslations { get; set; }
+        /// <summary>
+        /// Gets the choice translations
+        /// </summary>
+        public Dictionary<string, ChoiceTranslation> ChoiceTranslations { get; set; }
 
-        public Dictionary<string, AutocompleteChoiceTranslation> AutocompleteChoicesTranslations { get; set; }
+        /// <summary>
+        /// Gets the auto complete choice translations
+        /// </summary>
+        public Dictionary<string, AutocompleteChoiceTranslation> AutocompleteChoiceTranslations { get; set; }
     }
 
     public class AutocompleteChoiceTranslation
     {
-        public string AutocompleteName { get; set; }
+        /// <summary>
+        /// Gets the autocomplete choice name.
+        /// </summary>
+        public string AutocompleteChoiceName { get; set; }
 
         public DiscordApplicationCommandLocalization AutocompleteNameLocalizations { get; set; }
 
@@ -99,6 +143,9 @@ namespace DisCatSharp.ApplicationCommands
 
     public class ChoiceTranslation
     {
+        /// <summary>
+        /// Gets the choice name
+        /// </summary>
         public string ChoiceName { get; set; }
 
         public DiscordApplicationCommandLocalization ChoiceNameLocalizations { get; set; }
