+++ DisCatSharp.ApplicationCommands/Entities/OptionTranslator.cs
@@ -26,15 +26,36 @@ using Newtonsoft.Json;
 
 namespace DisCatSharp.ApplicationCommands
 {
-    public class ChoiceTranslator
+    public class OptionTranslator
     {
+        /// <summary>
+        /// Gets the option name.
+        /// </summary>
         [JsonProperty("name")]
         public string Name { get; set; }
 
+        /// <summary>
+        /// Gets the option name translations.
+        /// </summary>
         [JsonProperty("name_translations")]
         internal Dictionary<string, string> NT { get; set; }
         [JsonIgnore]
         public DiscordApplicationCommandLocalization NameTranslations
             => new(this.NT);
+
+        /// <summary>
+        /// Gets the option description translations.
+        /// </summary>
+        [JsonProperty("description_translations")]
+        internal Dictionary<string, string> DT { get; set; }
+        [JsonIgnore]
+        public DiscordApplicationCommandLocalization DescriptionTranslations
+            => new(this.DT);
+
+        /// <summary>
+        /// Gets the choice translators, if applicable.
+        /// </summary>
+        [JsonProperty("choices")]
+        public List<ChoiceTranslator> Choices { get; set; }
     }
 }
