+++ DisCatSharp.ApplicationCommands/Entities/SubGroupTranslator.cs
@@ -22,32 +22,43 @@
 
 using System.Collections.Generic;
 using DisCatSharp.Entities;
-using DisCatSharp.Enums;
 using Newtonsoft.Json;
 
 namespace DisCatSharp.ApplicationCommands
 {
-    internal class CommandTranslator
+    /// <summary>
+    /// Represents a sub group translator.
+    /// </summary>
+    internal class SubGroupTranslator
     {
+        /// <summary>
+        /// Gets the sub group name.
+        /// </summary>
         [JsonProperty("name")]
         public string Name { get; set; }
 
-        [JsonProperty("type")]
-        public ApplicationCommandType Type { get; set; }
-
+        /// <summary>
+        /// Gets the sub group name translations.
+        /// </summary>
         [JsonProperty("name_translations")]
         internal Dictionary<string, string> NT { get; set; }
         [JsonIgnore]
         public DiscordApplicationCommandLocalization NameTranslations
             => new(this.NT);
 
+        /// <summary>
+        /// Gets the sub group description translations.
+        /// </summary>
         [JsonProperty("description_translations")]
         internal Dictionary<string, string> DT { get; set; }
         [JsonIgnore]
         public DiscordApplicationCommandLocalization DescriptionTranslations
             => new(this.DT);
 
-        [JsonProperty("options")]
-        public List<OptionTranslator> Options { get; set; }
+        /// <summary>
+        /// Gets the command translators.
+        /// </summary>
+        [JsonProperty("commands")]
+        public List<CommandTranslator> Commands { get; set; }
     }
 }
